Optics Rx Equalizer setting to be done from Module state machine Discovered state
Summary:
We were trying to apply the RxEqualizer setting to optics from the code where we set the Application Select (ApSel). But for some of the port it does not work because if the module comes up with the default ApSel which is same as desired one then the code path gets skipped.

We want to apply the RxEqualizer from Module State Machine as it will be more predictable. The Equalizer setting depends on speed also and the speed value comes from Agent when port status is synced. Too we are applying it from Module Active state entry function.

Differential Revision: D28312306

fbshipit-source-id: fc5c9f9fe29e80e1543e38b58c9340b072eba17b