recursively pop and handle labels
Summary:
as titled.

if packet punted to cpu is programmed for pop and look up, simply pop label and handle with new MPLS header.

if label is known, check if label is programmed to cpu, if so handle label directly

if label is known, check if label is pop and look up. if so pop the label from stack (MPLS header is updated) and handle resulting header.

the recursion terminates at the bottom of stack, or if label is programmed to cpu or if label is unknown.

TODO: code to create and schedule IP packet.

Reviewed By: jasmeetbagga

Differential Revision: D28336104

fbshipit-source-id: 287748994f0176e5582167bc7c9c27b2e0dfc98b