Bring back optimization for only ser unreolved routes in rib
Summary:
As titled. We now have the ability to reconstruct
resolved routes from FIB (since its the same object that is
shared b.w RIB and FIB), so only serialize unresolved routes.
To land this only once the above optimization is already
in prod wedge_agent (so rollback is safe)

Differential Revision: D27560443

fbshipit-source-id: 53bec85468617c11f30a91501a704492033f5001