Verify in tests that diagCmds we run in FBAR are indeed safe
Summary:
We often need to run certain shell commands in
case of a outage to collect info for a later debug. Verify
that these are indeed safe. Note that in the test these
are run before verifying other invariants - copp, LB,
QOS mappings. This is done so if the cmds have any adverse
effects on the data plane, we shall see that in test failures.

TODO: move these safe cmds to a txt file that fbar code
can read as well.

Differential Revision: D27985030

fbshipit-source-id: 88ae4479432a7d02ebb7bb22940049635b38cfbc