Introduce new XphyMap to replace current 3D externalPhyMap_
Summary:
Since we're planning to allow PhyManager to use SW PortID to change
the xphy config for a FBOSS port, and current PlatformMapping has this
PortID : XPHY chip = 1 : 1 relationship, we should consider to use a
simplified map (2D vs 3D externalPhyMap_).

Besides there's no strong use case we need to deal with all the ports from
the same MDIO.

This design is also easier to let user to create one random XPHY for the
system(i.e. hw_test) without using vectors like externalPhyMap_ and
being cautious about the order of the mdio and phy.

Technically we can just use 1D map to store all the XPHY as we have
a unique xphy id for each one. But because we might need some pim-level
operations, like initializing all phys in the same slot(initializeSlotPhys)
or supporting multi-threading per pim. So we divide the whole xphy map based on PimID

  using PimXphyMap = std::map<GlobalXphyID, std::unique_ptr<phy::ExternalPhy>>;
  using XphyMap = std::map<PimID, PimXphyMap>;
  XphyMap xphyMap_;

Reviewed By: jasmeetbagga

Differential Revision: D28240365

fbshipit-source-id: e652e91977ef346cc521eb55496377efe9933174