Normalization: support adding entity prefix for normalized counters
Summary:
We add an entity prefix to the normalized counters which allows us to distinguish between prod vs. testing data.
By default, This will add a "test_" prefix to all counters so that we can rest assured that these counters will not conflict with the actual counters in prod.
We will need to set the prefix to "" for counters published to prod.

Differential Revision: D28369970

fbshipit-source-id: 81f9101d6d401c85b97c69d11185be28c87e03d4