Fix weights in show route output for large weight groups
Summary: For capacity aware routing use case, the nexthop weights will be large values - P413340813. So agent reduces the weights by proportionally reducing the weights. If the reduced total weight is greater than 128, we pad it to 512 in the hardware layer. However the show command will not display right weights because in RouteNexthop code we expect either a proportional reduction in weight or padding to 512 but not both. This assumption was right for turbo use case where total width will be <=512 always but with CAR, the large weight values means the assumption is not true

Reviewed By: saifhhasan

Differential Revision: D28274254

fbshipit-source-id: e707981af7d91283d34cfa3775e0feebeba53412