Remove extra bcm_pktio_init to fix PKTIO warm-boot
Summary:
With the two SDK patches (SDK-251784-6_5_17.patch and
SDK-254741-6_5_22.patch), the interal credit issue has been fixed.
However, the warm-boot was still failing because of one extra
bcm_pktio_init() from agent.  The doc said this call needed to be
made.  But it is actually already called from SDK.  Calling it again
cause the cmicx_pdma_credits_release() be called in the cold-boot
way, corrupting the warm-boot state.

With it removed and with the two SDK patches, warm-boot from
non-PKTIO to PKTIO is now working.

Reviewed By: msomasundaran

Differential Revision: D28499751

fbshipit-source-id: 665cff9a29338c1189ed0e33ebd67d1ef589abc8