Refer to agent config when configuring ports for a particular profile
Summary:
As titled. Look at the agent config and setup ports of
a particular profile (as desired by the test).

TODO:
- I had to put in a hack to configure ports on xphys only since
I ran into a problem on elbert where - ports were enabled agent config
but I could not find a xphy for it. Looking for inputs on refining this
- Expand tests beyond Elbert, first target will be other XPHY platforms
YAMP, Minipack

Reviewed By: birdsoup

Differential Revision: D28621507

fbshipit-source-id: 9ee870541614eab3e14441e069f3a85d83600004