Support PhyManager::programOnePort with TransceiverInfo
Summary:
Based on what we have on ExternalPhyPort<PlatformT, PortStatsT>::portChanged()
We need to use TransceiverInfo to generate the platform config to program xphy.

Therefore, adding an optional input for PhyManager::programOnePort() so that
we can support other than Elbert platform to implement such function.
This is one of the efforts to migrate current xphy programming from wedge_agent to qsfp_service.

https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/fboss/agent/platforms/wedge/ExternalPhyPort.cpp?commit=941b9880c49be54501039a712259096f576b6288&lines=47-84

Reviewed By: jasmeetbagga

Differential Revision: D28811590

fbshipit-source-id: e05b928703f71578456c8d564e186bc030e910bd