Fix first remediation interval after port sync happens
Summary:
This is the second issue w.r.t remediation.
When the system comes up, we make the lastDownTime_ as “now-4min” so that first remediation can happen in 2 minutes (remediation interval is 6 mins). But once the qsfp_service starts then Agent syncs up port information to it and the lastDownTime_ is set to.”now” which causes first remediation to happen after 6 mins.

To fix this when the Agent is syncing up the information then the lastDownTime_ is set to “now-4min” so that the first remediation could still happen after 2 minutes.

Reviewed By: birdsoup

Differential Revision: D28842033

fbshipit-source-id: 72837e62c6b1a0ea2ca5b4b0640dd8726eb9df9a