Introduce models for show transceiver
Summary: show transceiver defaults to some port information (id, status, etc.) in the cases when transceiver is not present from qsfp service. I think this is a good time to try to come up with a good way to display multiple types of objects. This is just an example of how this can be done, I can iterate on it if there are concerns. Alternatively we can discuss if we want to even allow something like this.

Reviewed By: hegara

Differential Revision: D28269442

fbshipit-source-id: 23b835e52d5c308ce67b2aa23dcce697e3d230d9