add sampled packet trap log to help debug S230947
Summary:
When debugging S230947, we saw a lot of packets were punt to CPU during SEV, but
unfortunately not reproducible later. To help debug similar issues in the future, I added
sampled logging for trapped packets at a higher logging level, so as to understand what and
why packets were punt to CPU if similar issue happens again.

Differential Revision: D28849071

fbshipit-source-id: ee885cad2724c0c12a641bad012f3c93899585a5