Clean up old googletest conditional compilations
Summary:
Remove all the conditional compilations that were used for the googletest upgrade. We're set on Googletest 1.10.x now, so no need to keep the old code lying around.

TODO in the future, is to deprecate some of the other fb-only patches, and to switch over fbcode to use fbsource/third-party

Reviewed By: yfeldblum, igorsugak

Differential Revision: D28860854

fbshipit-source-id: 8ffbaf860fca8897f6691218bc0154723ec954d4