Support CheckDefaultXphyFirmwareVersion hw_test
Summary:
The way we used std::optional attribute might be an issue that we never actually assign the PhyFwVersion to this optional attribute.
Which will cause the test still thinks such optional attribute is not set, so it will skip the test.

Since we know which platforms have xphy, we can just create a PhyFwVersion object and assign version info for those platforms.
As for those who don't have XPHY, we can just return in this test.

I also added a toString for PlatformMode for pretty logging.

Reviewed By: jasmeetbagga

Differential Revision: D28851847

fbshipit-source-id: 39fd1c7d452db10307f70654a3421ef76eaf912b