Route stat bcm cmd support in Cinter and Fake SDK
Summary:
This is the first diff in a stack of diffs to add route counter id support. Design document - https://fb.quip.com/6W1mAEblcVM7. This diff stack adds the ability to attach a counter id to routes.
TODO:
- Route scale convergence tests.

Changes in this diff:

Adding cinter and fake sdk support for APIs that are needed for route counter support.

Differential Revision: D28904459

fbshipit-source-id: aeace870c039bc1cee82730df65305f2e37edce8