Support multi-threading initializng pim xphy
Summary:
Similar to what we have in wedge_agent, Pim.cpp:
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/fboss/agent/platforms/wedge/facebook/Pim.h?commit=762bd46bd10a&lines=43-44

each Pim of MultiPimPlatform has a pimThread_ and pimEventBase_ so that when we try to initialize all pims of a multi-pim platform,
we can use the EventBase from each pim to initialize all pims in parallel.
https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/fboss/agent/platforms/wedge/facebook/MultiPimPlatform-defs.h?commit=762bd46bd10a&lines=19-34

As we decided to move the whole xphy programming from wedge_agent to qsfp_service,
we should also move this multi-threading mechanism to current PhyManager.

Therefore, I introduced PhyManager::setupPimEventMultiThreading(PimID) and call it during initExternalPhyMap()
when we create the xphy for each valid pim. This function will create a `PimEventMultiThreading` unique_ptr for such pim
to manage the pimThread and pimEventBase.

And then in WedgeManager::initExternalPhyMap(), we'll try to fetch the pim EventBase* if there's one.
Note that, for ElbertPhyManager, we might need to deal with mixed 16Q and 8DD pims, and only the latter would need to program xphy.
If there's an pim EventBase, we'll use it to create a folly::Future for such pim initializing xphy.
Otherwise, we will just call the initialization function directly.

Reviewed By: jasmeetbagga

Differential Revision: D29122288

fbshipit-source-id: 8828bdcfaad5963858b5997b8c087b24f81159d6