Sai Replayer automatic object type/name detection
Summary:
This is towards supporting Sai Attribute names in Sai Replayer. It also skips the attribute type detection, such that adding new attributes will be as simple as

  SAI_ATTR_MAP(AclTable, FieldEthertype)

The set*Attribute() function is also generalized to remove the redundant works. This is achieved by keeping a map of attribute id -> (attribute name, attribute type hash code). The following code change is just example of AclTable and AclCounter.

Will need to fix some naming of attributes to follow the Sai spec.

Reviewed By: jasmeetbagga

Differential Revision: D28941231

fbshipit-source-id: acee9a12c1a9b4e0233f3a47d0fd7de7b189accd