Running transceiverPortsChanged in i2c event base thread - version 2
Summary:
This is the second attempt to fix the the deadlock issue we are seeing on some systems with CMIS modules.
The issue is that the qsfp module level lock is acquired by multiple threads. The refresh loop running under the I2C evb thread acquires the module level lock. When syncPorts comes, it calls transceiverPortsChanged which also acquires the module level lock but in the thrift handler thread. This causes the event base thread to hang waiting for the module level lock sometimes.

The first attempt to fix is in D29123005. The difference between two versions is the parallelization. The first fix can make transceiverPortsChanged run in parallel on systems with multiple i2c controllers. This version keeps the original behavior of running them sequentially. I will introduce parallelism in another diff in an attempt to keep this diff simple and unblock others.

Reviewed By: jasmeetbagga, rajank7

Differential Revision: D29258511

fbshipit-source-id: 26a6651b51917d0240b1e5ec35c997783382b73d