Assert watermarks for queues on more than one port
Summary:
As titled. Recently we introduced a bug whereby
watermarks were reported only for queues on the first port
https://fb.workplace.com/groups/668135506568401/posts/5932446560137243

This escaped our tests since the tests were using a config with
only one active port. Change that to use multiple ports now, and
assert for watermarks for 2 queues on 2 ports

Reviewed By: shri-khare

Differential Revision: D29288772

fbshipit-source-id: df636dc885b305c828555ffd58fb7ec1887698f0