Add device_frequency config for leaba
Summary:
Currently, agent config does not explictly specify the
device_frequency to use, which results in the default SDK value of
1200000 being picked up. This could result in unexpected behavior as
leaba chip is actually using a frequency 1350000 Khz. Adding an option in
config to explicitly specify this.

gkurio thanks for confirming 1.35GHz is the correct frequency for the core VCO in T93174991.

Reviewed By: jasmeetbagga

Differential Revision: D29252280

fbshipit-source-id: f8971cbd0edb8b42056518c5197848f4be7f94f4