Fix the Config reject due to bad ApSel lanes issue
Summary:
This diff fixes the issue where in some Finisar optics the Rx Equalizer setting gets rejected with error as "Invalid ApSel selected". We found that the 200G mode Finisar optics contains invalid values on last 4 lanes ApSel register. If RxEqualizer function tries to apply config for all 8 ports in 200G mode then the module rejects it because the last 4 lanes don't have valid config.
To fix this we need to apply config to only 4 lanes in 200G/100G mode and 8 lanes in 400G mode.

Differential Revision: D29208521

fbshipit-source-id: aa8a3f2113150f966cf4c70f1ce547bd02ea5566