Cache LaneList for xphy port
Summary:
Based on current ExternalPhy design, it's hard to get which lanes that
a SW port is using. Because all the ExternalPhy are using lanes to program
the configs directly instead of a software port.

Since we always use PhyManager to programOnePort(), we can cache the
LaneID list of both system and line sides in PhyManager. And then for
all the following xphy related logic like, programming prbs, getting stats,
we can just use this cached info to pass in the xphy lanes directly.

NOTE: This cache is kinda like wedge_agent SwitchState that we might need to consider serialize it during qsfp_service shut down.
When it warmboot, we can recover it from the cache files.
I will add such logic in the later diff.

Reviewed By: rajank7

Differential Revision: D29398763

fbshipit-source-id: 37be92d416f5382b9ac3fcacc3eac73f10df90d0