Added MYIP CoPP invariant test
Summary:
In the agent config, an invariant already existed for BGP; the purpose of this task was to add another invariant. I added an invariant test for MYIP by adding a function to the test class which pulls out the `sendPkts()` lambda into a separate function which takes the destination port and queue ID as arguments - this allows us to implicitly specify which CoPP to test.

Specifically, BGP packets go to the high-pri queue and MYIP packets to the mid-pri queue.

Differential Revision: D29407112

fbshipit-source-id: 2a68506ea88ae01f2df35ec01e7c88abb1e63f9d