HwTests for interaction b/w my station and l2 tables.
Summary:
Tests added in response to S234435. We hit a vendor SDK bug,
whereby when router MACs aged out on a particular VLAN, the
MAC got pruned  from MY STATION table in addition to the L2
table. This then broke routing on ports belonging to such a VLAN
The test below induces MAC learning and aging of router MAC
on ALL vlans, ingresses a packet on one such port, and asserts
that the packet was actually routed correctly

Differential Revision: D29285241

fbshipit-source-id: 5a758781bee6ef693e8ccfaef4ad3355f10a12c1