Publish qsfp_service's ODS counters with correct port names
Summary:
This is the first change to start publishing ODS counters with the correct port/interface named keys. Currently the keys are named after the transceiverID which is misleading on modular systems. For example, the current keys are named like ```qsfp_service.qsfp.port.eth1/0/1.present```. This name doesn't make sense on a minipack.

In the new method, the same counter will be named like
```qsfp_service.qsfp.interface.eth2/1/1.present```

We will need to support the old keys for some time until everyone can transition to the new ones, therefore in this change I start duplicating two of the counters, one with the old key, and one with the new key named after interface name.
In future diffs, I'll duplicate rest of the counters also

Reviewed By: rajank7

Differential Revision: D29450375

fbshipit-source-id: 24da054aaeec92c84792488d076822cdd42e771c