Delete QsfpModules on hard reset
Summary:
We should delete QsfpModule(Sff/CmisModule) objects on triggerQsfpHardReset to remove any previous cached state. This isn't really a prod problem because if someone removes a module in production, the next refresh will detect that the module is absent and delete the QsfpModule object.

However, we use the triggerQsfpHardReset() in qsfp tests, and without this change the QsfpModule object ( and it's old state ) stick around after reset, causing the tests to require multiple refresh cycles before the module has expected settings.

Reviewed By: jasmeetbagga

Differential Revision: D29540402

fbshipit-source-id: 979c5fe12c52cedab4052e6ff090e32fb3e7947e