Document ClientIDs and define AdminDistance for each ClientID
Summary:
1. Document what each ClientID refers to
2. Explicitly define priority for each ClientID. STATIC_ROUTE, INTERFACE_ROUTE,
   LINKLOCAL_ROUTE, and STATIC_INTERNAL were missing.
3. Priority for missing ClientID is borrowed from Agent code.
4. With (3), we will avoid mis-leading log of ClientID resolution failure - https://fburl.com/code/qutmhhua

With (2), we can rely on `Configuration` for priority specification of every
ClientID. We may no longer need to explicitly specify priority for every route
with generate. In future we can easily tweak priorities e.g. Introducing new
ClientID with priority between CONNECTED_ROUTE and STATIC_ROUTE will be
challenging as of current implementation but would be lot easier to do if we're
relying on configuration.

Reviewed By: GirasoleY

Differential Revision: D29602294

fbshipit-source-id: 70264a4a681ad350d44d5b61247987aebf95f93f