Fix --vdm_info cli for host BER in optics
Summary:
The optics have the host BER info in page 0x21 and 0x25 at location 32 onwards. In page 0x25 the offset 0 to 31 are empty so the CLI to dump vdm-info was not parsing this page assuming no valid data to be present.

Changed the logic to look for vdm-info all across the page and not assume no more valid data on seeing 0x00 at a pair of bytes.

Reviewed By: birdsoup

Differential Revision: D29607394

fbshipit-source-id: 64f9ed78c020e6136de2870ac0922199efdd9a0e