Support warm/cold boot notion in qsfp tests
Summary:
Idea here is to test that qsfp WB (normal course
of operation), comes back with all the transceiver and xphy
settings intact. So in tests we will have a notion of running
a verify post WB. While on CB we will still run both setup + verify

TODO: Cold boot on qsfp service actually fails port profile
tests since it resets transceivers and it seems they don't recover
in time for the test to program. Fix that. Will land only after
we have a fix.

Differential Revision: D29383337

fbshipit-source-id: 7c04f80a0f8a5361250990d0995a7884ffe5c6af