T90111204: Add optics remediation counter to ODS for FBOSS switches
Summary:
https://www.internalfb.com/tasks?p_nav=SUBSCRIBED&t=90111204

Add a counter named remediationCounter to transceiver.thrift to keep track of the number of times the optics remediation is performed on a given port. This counter will be updated to ODS at a select interval in method StatsPublisher::publishStats(). Method QsfpModule::refreshLocked() will increment this counter whenever remediation is performed.

Differential Revision: D29046836

fbshipit-source-id: 58a608fba359054b03d6d03025ae1b13b5771afb