Move MDIO base address to container
Summary:
FbFpgaRegister.h is used by both Minipack and CR. The MDIO command register
addresses to day contain the Minipack base address as well. So, we cannot use
them as is for CR.

Here we are moving the base address to Minipacck container, so that CR can
re-use the same code.

Reviewed By: jasmeetbagga

Differential Revision: D29777943

fbshipit-source-id: 01274800edde6ef88d384bc19df4d2124d011a37