Add HwInvariantBitmask type to pass test option to switch roles
Summary:
Instead of creating a slew of new boolean flags to pass into functions, we define a bit-mask datatype with enumerated values to be able to set and check multiple options with a single parameter.

A convenience function for checking whether a bit is set is also added, and we start out with the underlying datatype as a 32-bit unsigned integer. If more options are eventually desired, we can change the typedef to be a 64-bit int at any point.

Differential Revision: D29658394

fbshipit-source-id: 9b6c4725b317dc980e6614d93c0080c48a3724be