Add test to check for cpu queue stuck case
Summary:
Add a new test to check for CPU queue stuck condition as seen in issue
reported in wdg400c-331. The issue happens when there is congestion in a queue
configured with ECN profile, hence configuring ECN on CPU queues and congesting
the same to run into the issue.

- Add option to configure ECN for CPU queues.
- With ECN configured on CPU queues, unable to reach line rate and hence relaxing the check to be able to reach line rate.
- Send high rate traffic which will congest CPU queue, causing CPU stuck.

The test seems to fail for th/th3 platforms for SAI, as setting ecn threshold is not something we'd be interested in, wondering if we still need to get this fixed from Broadcom via a CSP to make sure this test work, or should we have this tested only for TAJO?

Reviewed By: jasmeetbagga

Differential Revision: D29415831

fbshipit-source-id: b22ef25061a0ecc350ece45cbb9ea6630c494e22