Support store warmboot qsfp_service_state to disk
Summary:
Because most of the ExternalPhy chips don't have the logic to restore the system or line side
lane id vector based on a port, during warmboot we will lose the mapping b/w a software port id
and the lane info.
Therefore, it will fail current warmboot qsfp_hw_test since we don't program the port again during
qsfp_service restart in warmboot mode.

To fix that, we need to adapt the same warmboot cache mechanism from wedge_agent in qsfp_service.

The first step is to store the qsfp_service state in the warmboot cached file in the disk.

Reviewed By: birdsoup

Differential Revision: D29861866

fbshipit-source-id: 1876a1ac2f04be46dff23ae6bb8f1fde83873001