Test cabled ports only in Port Profile Tests
Summary: as titled. There are certain ports that have optics plugged in but are disabled in the agent config because there is something wrong with them. Those ports have an unknown profileID which can make the test fail. The HwPortProfileTests should just run on ports that we know are cabled.

Reviewed By: birdsoup

Differential Revision: D29870802

fbshipit-source-id: 20fb41a11619fe9472c36caae8ddf65cc9cedc92