Test to assert that warm boot is hitless on link tests
Summary:
- Create a data plane loop, via the familiar technique
of using router mac as next hop mac and disabling TTL decrements.
Now each packet sent to the configured route, will loop to the
other cabled port, only to be routed again endlessly.
This is sufficient to create a pretty high rate of traffic.
- Assert that no in discards are seen on WB here. If there
are port flaps, we would end up hitting in discards.

TODO:
- Configure LB in config so traffic is distributed evenly and
all ports are see sufficient traffic to detect in discards on
flap

Depends on : D29921010

Differential Revision: D29925381

fbshipit-source-id: d101e8d91fba3232ef7434f82a4c4efa7d6e6679