Assert more in warmboot is hitless test
Summary:
While in discards is a good determinant of loss,
ECMP width is a even better one. Depending on how fast we
shrink the ecmp group and the traffic rate, we may just fall
under the in discard threshold. Assert for ecmp width as well,
since if a port flaps during WB we will shrink ecmp and never
expand it.

Differential Revision: D29927085

fbshipit-source-id: ccd0027c4561976dd96839618b78b63a80b94c9a