Skip HW write when we just need to update sai object memory
Summary:
For route metadata, there is a case where we need to
update just our sai object memory with the HW value, before updating the HW.
Since HW already has this value, skip the extra HW write. Besides being
a bit more efficient, this also avoids a problem in brcm-sai, viz.
- BRCM-SAI uses special metdata values for to cpu routes, and prohibts
using these values. With the existing logic, its possible that we
read one of these special values and try to set them in HW. Ideally the
SAI adaptor should check that the value is not really being changed, and
make it a noop, but the check for value being prohibited comes first.
So handle in our code

Reviewed By: shri-khare

Differential Revision: D29966363

fbshipit-source-id: b89c76640c6bebaec799726c9bf2878a9658c713