Set SWITCH_ATTR_INGRESS_ACL on ACL Table Group
Summary:
Previously, this wasn't supported, but 5.0.0.7 onward this is supported.
Settig this on 4.2.2.9 is no-op, and warmbooting from 4.2.2.9 to 5.0.0.7 with
this set is OK too. So OK to set this uncodtionally.

We need to audit Features that are always True on all platforms and remove as
applicable.

Would explore that in follow up stacked diff for INGRESS ACL Feature as well as
other features.

Reviewed By: zechengh09

Differential Revision: D29954688

fbshipit-source-id: 1f6e1f4a0da2374fccbfef6c2e094c1177243833