Incorporate option checks into InvariantHelper verifyInvariants(), extend existing InvariantTests to include options
Summary:
With the HwInvariantBitmask datatype defined, we can now incorporate these option checks into the `verifyInvariants()` function.

In ProdInvariantsHelper, these fields are checked in `verifyInvariants` before actually running the tests.

In ProdInvariantsTest, a virtual `getInvariantOptions()` function is introduced to get the correct options for each switch role. `verifyInvariants` is also modified to accept an options parameter.

We also make it an error to pass in no invariant options into ProdInvariantHelper's `verifyInvariants`, since we always want to test something. As such, other places that call this function have been modified to pass in a bitmask that represents the functionality prior to our changes.

Differential Revision: D29658391

fbshipit-source-id: eed85efb33b3686366b030448cbe3461eb296adb