Clear and set port serdes attributes on TH3
Summary:
Issue: Brcm-sai requires setting the all sixtap attributes at once, while sai spec only supports setting one attribute at a time. This is a bug in the brcm-sai implementation, but we'll need a workaround on this.

Solution: Instead of setting attributes, we remove the port serdes and then recreate the serdes objects with sixtap attributes. The newly created serdes object would have the same object Id, since it's tied to the same port.

Reviewed By: shri-khare

Differential Revision: D29442757

fbshipit-source-id: 842e6a36f8d0f4de205b652aea93c91e145d002c