Helper function to get number of uplinks from HwSwitch* argument
Summary:
When creating prod-setting configs, we start with a fresh config created through ConfigFactory's `createUplinkDownlinkConfig`, which takes in the number of uplinks in the switch as an argument.

To supply this argument, I've hardcoded the uplink values from `getUplinksCount` in HwInitAndExitBenchmarkHelper.cpp in a switch case, to get the value from a given HwSwitch pointer.

Differential Revision: D29921490

fbshipit-source-id: b05f42185b3ea8484329d230087ba57552e28d30