Add tests for counter ids with drop and punt routes
Summary:
Add tests to verify counter ID handling for drop and punt packets.

Note:
At hw layer broadcom uses 2 separate counter ids for forwarded and dropped packets. To reduce the number of counters used, we haven't enabled drop counting in bcm hw. This test verifies the functionality at RIB layer so that this can be used on other asics if they support counting drops with 1 stat id (to confirm) or to enable drop counting with bcm at a future point

Reviewed By: jasmeetbagga

Differential Revision: D29978843

fbshipit-source-id: 04a5a7cda7c10af138703a90de2c1995b51ff7c2