Refactor to make getMMUCellBytes more generic.
Summary:
(1) Today, getMMUCellBytes only apply to BCM platform. Pull that up t be more generic for SAI and BCM platforms. Need this abstraction to be able to control getMMUCellBytes from HW Tests.
getMMUCellBytes is not implemented for platforms other than BCM for now, but an error will be thrown if we try to use it.
(2) Expose getPgMinLimitBytes as public interface for tests in subsequent diffs.

Reviewed By: msomasundaran

Differential Revision: D29997462

fbshipit-source-id: f39e9ce14402fab2a9b7643c679bf52756ab3b01