Trigger an initial refresh in the setUp of qsfp hw tests
Summary: as titled. Triggering a refresh forces the optics to customize based on the speed in the agent config before the actual test starts. We also need a 5 second delay after the customization happens otherwise we see that certain CMIS modules are still initializing(changing speed) when the test starts. This fixes the HwTransceiverReset problem seen on CMIS modules on yamp

Reviewed By: rajank7

Differential Revision: D29929238

fbshipit-source-id: 4a21371ac01456d974575a25d3524bf980f6fef9