Event generation for module state machine should be protected for multi-threading
Summary:
The problem is that the event generation for state machine needs to be protected for multi-thread case. Two thread can’t generate the events at the same time. I earlier thought such things will be taken care by Boost MSM internally but looks like we need to make sure this.

In our code most of the place where the event gets generated, is protected by module lock. Only in case where the event gets generated by function scheduler, the lock is not acquired so that causes issue. Also the QsfpModule() destructor generates event which also needs to bee protected. With this change the test passes without data race.

Differential Revision: D29999721

fbshipit-source-id: cb13df085d8e10c0d303c169bad35ca2aecb4649