Fix numHostLanes and numMediaLanes
Summary: These APIs currently return the private members numHostLanes_ and numMediaLanes_ which only get updated when an application code is changed. The application code doesn't change on warmboots and in that case numHostLanes and numMediaLanes return the default 4 always. These values are not true for 400G modules which have 8 host lanes (and 4 media lanes).

Reviewed By: rajank7

Differential Revision: D30126182

fbshipit-source-id: e829b4150481a3608570364cc6dad9bb34cce639