Pick different priorities for peers to allow for keyserver election
Summary:
As titled. Without this, we can't elect keyserver and generate
SAK (session association key). Would like to hear from reviewers
on how this will be handled in prod. Since we depend on EBB
keyserver to program these keys, do we expect peer priority
to also come as part of that call?

Reviewed By: rajank7

Differential Revision: D30121493

fbshipit-source-id: dc229663f76280f80d6d4529e7a4fd4bd3a5e4c5