Fix the T2 specific invariant tests.
Summary:
This one fixes couple of issues with the invariant tests for T2 asic.
(1) 100G speed explicitly set as port speed is rejected by T2. Assign 40G instead for T2
(2) OLYMPIC_QOS_INVARIANT should not be validated on T2 as it doesn't support this feature. When setting  invariants to validate, explicitly check for the support of the feature.

Reviewed By: rsunkad

Differential Revision: D30180550

fbshipit-source-id: 2f94407447a1046b074143eaad882aa3b2ca4bc3