Keep track of the named queues per port
Summary:
Add a mechanism to keep track of the named queues for a port. Currently,
this information is not readily available associated with queues, but is
available in counters and its not easy for use when iterating over queues.
Making this easily available for use during every BST stats poll, where we need
to just publish/update stats for enabled queues and for CPU queues, the enabled
queues are the only ones with a name associated with them.

Reviewed By: jasmeetbagga

Differential Revision: D30023742

fbshipit-source-id: 7a4ed8c2cf24e42ccd17678a0192b3cb047575b1