Work with cabledPorts only in HwTransceiverResetTest
Summary:
Currently this test fails on elbert on a blanco module. Blanco is a passive module and doesn't support any monitors like the rest of CMIS modules do. Working with cabledPorts fixes the test.
Also, in general it's better to only verify cabledPorts in this test. The present but non cabled ports could be bad transceivers and thus marked as not cabled.

Reviewed By: jasmeetbagga

Differential Revision: D30213730

fbshipit-source-id: 82a1f1c808ef5c31cd189d8f676f77dd46e926b9