link down event signal slot
Summary:
In the current model of link handling, in response to link down event, FDB entry is reset. This triggers a chain of reset, leading to deletion of neighbor, nexthop and next hop group member causing shrink of ECMP group. This shouldn't be a problem, since link down event typically causes NDP or ARP flush which eliminates FDB entry. This however is specific to only static entry and not dynamic entry which are also flushed in response to link down. Secondly, while FDB entry is flushed from SAI store (and HW), managed FDB entry still survives which never recovers because its publisher objects (such as bridge port or router interface) never undergo reset cycle (i.e. never deleted and recreated). This leads to mismatch in SW SwitchState and HW Switch state. The linkdown slot is being proposed to mitigate that.

Let SwSwitch schedules delta in response to link down to purge MAC entries and neighbor entries, but link down only causes fast shrink.

Further in BCM Switch, link down only triggres reprogramming of host entry to CPU and shrink of ECMP group. This can also be seen as an attempt to bring SAI Switch closer to BCM Switch. The only gap currently is how to reprogam host entry to CPU? Next Hop objects are not programmed to CPU but are removed. Neighbor entries are also not programmed to CPU.

publisher object to notify link down
subscriber object to get notified of link down

Reviewed By: shri-khare

Differential Revision: D30195307

fbshipit-source-id: 02b07398eec90ee9ff7c3838f3daf92805ad9996