fdb publisher object to trigger link down event
Summary:
as titled. this is done as an alternative to current approach, to reset fdb entry,

on link down, instead of resetting SAI object, simply notify subscribers of link down event. An object which gets reset on link down is next hop.

This brings SAI switch implementation a bit closer to BCM Switch (where L2 entries are not flushed on link down).

This also puts Software Switch state in sync with that of SDK, earlier SwSwitch state had Mac Entry, but SAI switch had reset it on link down.

For more context please check D30195307

Reviewed By: shri-khare

Differential Revision: D30201709

fbshipit-source-id: b96c5ff2d12ddf35dd0e2dbea06c4f9da2f18187