Make PortCacheInfo synchronized in PhyManager
Summary:
Per offline discussion w/ jasmeetbagga, we need add some lock protection for PortCacheInfo in PhyManager.

Even though `portToCacheInfo_` in PhyManager won't have erase() operation because we always put all ports in the PlatformMapping
to this map in PhyManager constructor.
We might have scenarios which will change the `systemLanes` and `lineLanes` because the speed change of some ports,
like the controlling ports can subsumed all the lanes of subsumed ports.
Therefore the two attributes can change during the whole lifetime of PhyManager.
We need some lock protection especially some of the functions accessing this PortCacheInfo via multi-threading, like stats collection.

Thus in this diff, we'll change the existing portToCacheInfo_ to
```
  std::
      unordered_map<PortID, std::unique_ptr<folly::Synchronized<PortCacheInfo>>>
          portToCacheInfo_;
```

And provided a bunch of locked helper functions with LockPtr as an input.

Reviewed By: jasmeetbagga

Differential Revision: D30115945

fbshipit-source-id: a6dd34d3f012a3fa8a056aa2371acc16b90fec0b