Add queue-per-host verification functions to HwProdInvariantHelper
Summary:
Before we define a MHNIC class to test with, we must first define the actual verification functionality for queue-per-host mappings.

To accomplish this, we refactor the existing `verifyHelper()` function from HwQueuePerHostRouteTest.cpp into a generic `verifyQueuePerHostMapping()` function in HwTestQueuePerHostUtils.

This is then called from both ProdInvariantHelper's `verifyQueuePerHostMapping()` and HwQueuePerHostRouteTest's `verifyHelper()`.

Reviewed By: jasmeetbagga

Differential Revision: D30055474

fbshipit-source-id: d9545949e9fed7d89b8f01a4ce2de9d3156b9d84